// management-ui/src/app/models/employee.model.ts
export interface Employee {
    id?: string;
    firstName: string;
    lastName: string;
    avatarUrl: string;
    companyName: string;
    designation: string;
    email: string;
    contactNo: string;
    experience: string;
    joiningDate: string;
    team: string;
    manager: string;
}

export interface Navigation {
    name: string;
    icon: string;
    badgeCount?: string;
}